Split `matrices.py` into `matrices.py` and `common.py`
